split faker seed data into bcws/emcr
