--- backend/src/database/seed-functions.ts
@@ -1,15 +1,0 @@
-import { datasource } from './datasource';
-import { functionSql } from './queries';
-
-const seedFunction = async () => {
-  await datasource.initialize();
-  const functions = await datasource.query(
-    `SELECT * FROM public.emcr_function`,
-  );
-  if (functions.length === 0) {
-    return await datasource.query(functionSql);
-  }
-  return 'Already seeded';
-};
-
-seedFunction();
